Calculates average expression changes of all (or specified) nodes except trigger and finds the perturbed node count for all (or specified) nodes in system.

Calculates average expression changes of all (or specified) nodes except trigger and finds the perturbed node count for all (or specified) nodes in system.

find_node_perturbation(input_graph, how = 2, cycle = 1, limit = 0,
  fast = 0)

Arguments

input_graph
The graph object that was processed with priming_graph function.

how
The change of count (expression) of the given node in terms of fold change.

cycle
The iteration of simulation.

limit
The minimum fold change which can be taken into account for perturbation calculation on all nodes in terms of percentage.

fast
specifies percentage of affected target in target expression. For example, if fast = 1, the nodes that are affected from miRNA repression activity more than one percent of their expression is determined as subgraph.

Value

It gives a tibble form dataset that includes node names, perturbation efficiency and perturbed count of nodes.

Details

find_node_perturbation calculates mean expression changes of elements after the change in the network in terms of percentage. It also calculates the number of nodes that have expression changes after the change occur in the network. The outputs of the function are the perturbation efficiency and perturbed count of nodes for each nodes.

Examples

data('minsamp')
data('midsamp')

 minsamp%>%
  priming_graph(competing_count = Competing_expression, miRNA_count = miRNA_expression)%>%
  find_node_perturbation()%>%
  select(name, perturbation_efficiency, perturbed_count)

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tibble: 8 x 3
#>   name  perturbation_efficiency perturbed_count
#>   <chr>                   <dbl>           <dbl>
#> 1 Gene1                   0.272               3
#> 2 Gene2                   0.272               3
#> 3 Gene3                   0.153               3
#> 4 Gene4                   0.925               5
#> 5 Gene5                   0.381               2
#> 6 Gene6                   0.653               2
#> 7 Mir1                    1.63                4
#> 8 Mir2                    3.43                3

 minsamp%>%
  priming_graph(competing_count = Competing_expression, miRNA_count = miRNA_expression,
    aff_factor = c(energy,seed_type), deg_factor = region)%>%
  find_node_perturbation(how = 3, cycle = 4)%>%
  select(name, perturbation_efficiency, perturbed_count)

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tibble: 8 x 3
#>   name  perturbation_efficiency perturbed_count
#>   <chr>                   <dbl>           <dbl>
#> 1 Gene1                   0.203               5
#> 2 Gene2                   0.317               5
#> 3 Gene3                   0.101               5
#> 4 Gene4                   0.299               5
#> 5 Gene5                   0.269               5
#> 6 Gene6                   0.249               5
#> 7 Mir1                    1.60                6
#> 8 Mir2                    5.58                6

 midsamp%>%
  priming_graph(competing_count = Gene_expression, miRNA_count = miRNA_expression)%>%
  find_node_perturbation(how = 2, cycle= 3, limit=1, fast = 5)%>%
  select(name, perturbation_efficiency, perturbed_count)

#> Warning: First column is processed as competing and the second as miRNA.

#> Warning: `as_quosure()` requires an explicit environment as of rlang 0.3.0.
#> Please supply `env`.
#> This warning is displayed once per session.

#> Subsetting by edges

#> Warning: `cols` is now required.
#> Please use `cols = c(eff_count)`

#> # A tibble: 24 x 3
#>    name   perturbation_efficiency perturbed_count
#>    <chr>                    <dbl>           <dbl>
#>  1 Gene1                   NA                  NA
#>  2 Gene2                   NA                  NA
#>  3 Gene3                   NA                  NA
#>  4 Gene4                    0.525               5
#>  5 Gene5                   NA                  NA
#>  6 Gene6                    1.67                6
#>  7 Gene7                    0.386               0
#>  8 Gene8                    0.178               0
#>  9 Gene9                    0.806               6
#> 10 Gene10                   0.294               0
#> # … with 14 more rows