Utilizes the change in expression value/s as triggering.

simulate function uses the change in expression value/s as triggering.

simulate(input_graph, cycle = 1, threshold = 0, knockdown = TRUE)

Arguments

input_graph
The graph object that processed in previous steps.

cycle
Optimal iteration number for gaining steady-state.

threshold
absolute minimum amount of change required to be considered as up/down regulated element

knockdown
specifies gene knockdown with default TRUE

Value

The graph.

Details

The steady-state conditions of the system are disturbed after the change in the graph (with update_how or update_variables). In this case, the system tend to be steady state again. The arrangement of competetive profiles of the targets continue until all nodes are updated and steady-state nearly. Note that, If `how` argument is specified as `0`, *simulate()* and *update_how()* functions process the variables to knockdown of specified gene with default `knockdown = TRUE` and knocked down competing RNA is kept at zero. However, if `knockdown= FALSE` argument is applied, competing RNA which has initial expression level of zero is allowed to increase or fluctuate during calculations.

Examples

data("minsamp")
data("new_counts")

## new_counts, the dataset that includes the current counts of nodes.

priming_graph(minsamp, Competing_expression, miRNA_expression)%>%
  update_variables(new_counts)%>%
  simulate()

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tbl_graph: 8 nodes and 7 edges
#> #
#> # A rooted tree
#> #
#> # Node Data: 8 x 7 (active)
#>   name  type      node_id initial_count count_pre count_current changes_variable
#>   <chr> <chr>       <int>         <dbl>     <dbl>         <dbl> <chr>           
#> 1 Gene1 Competing       1         10000     10000        10063. Up              
#> 2 Gene2 Competing       2         10000     20000        19841. Down            
#> 3 Gene3 Competing       3          5000      5000         5032. Up              
#> 4 Gene4 Competing       4         10000     10000        10063. Up              
#> 5 Gene5 Competing       5          5000      5000         5000  Competing       
#> 6 Gene6 Competing       6         10000     10000        10000  Competing       
#> # … with 2 more rows
#> #
#> # Edge Data: 7 x 20
#>    from    to Competing_name miRNA_name Competing_expre… miRNA_expression dummy
#>   <int> <int> <chr>          <chr>                 <dbl>            <dbl> <dbl>
#> 1     1     7 Gene1          Mir1                  10000             1000     1
#> 2     2     7 Gene2          Mir1                  10000             1000     1
#> 3     3     7 Gene3          Mir1                   5000             1000     1
#> # … with 4 more rows, and 13 more variables: afff_factor <dbl>,
#> #   degg_factor <dbl>, comp_count_list <list>, comp_count_pre <dbl>,
#> #   comp_count_current <dbl>, mirna_count_list <list>, mirna_count_pre <dbl>,
#> #   mirna_count_current <dbl>, mirna_count_per_dep <dbl>, effect_current <dbl>,
#> #   effect_pre <dbl>, effect_list <list>, mirna_count_per_comp <dbl>

priming_graph(minsamp, Competing_expression, miRNA_expression,
       aff_factor = c(seed_type,energy), deg_factor = region)%>%
  update_variables(new_counts)%>%
  simulate(cycle = 3)

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tbl_graph: 8 nodes and 7 edges
#> #
#> # A rooted tree
#> #
#> # Node Data: 8 x 7 (active)
#>   name  type      node_id initial_count count_pre count_current changes_variable
#>   <chr> <chr>       <int>         <dbl>     <dbl>         <dbl> <chr>           
#> 1 Gene1 Competing       1         10000    10064.        10064. Up              
#> 2 Gene2 Competing       2         10000    19997.        19997. Down            
#> 3 Gene3 Competing       3          5000     5023.         5023. Down            
#> 4 Gene4 Competing       4         10000    10028.        10029. Up              
#> 5 Gene5 Competing       5          5000     5000.         5000. Down            
#> 6 Gene6 Competing       6         10000    10000.        10000. Down            
#> # … with 2 more rows
#> #
#> # Edge Data: 7 x 23
#>    from    to Competing_name miRNA_name Competing_expre… miRNA_expression
#>   <int> <int> <chr>          <chr>                 <dbl>            <dbl>
#> 1     1     7 Gene1          Mir1                  10000             1000
#> 2     2     7 Gene2          Mir1                  10000             1000
#> 3     3     7 Gene3          Mir1                   5000             1000
#> # … with 4 more rows, and 17 more variables: seed_type <dbl>, energy <dbl>,
#> #   region <dbl>, dummy <dbl>, afff_factor <dbl>, degg_factor <dbl>,
#> #   comp_count_list <list>, comp_count_pre <dbl>, comp_count_current <dbl>,
#> #   mirna_count_list <list>, mirna_count_pre <dbl>, mirna_count_current <dbl>,
#> #   mirna_count_per_dep <dbl>, effect_current <dbl>, effect_pre <dbl>,
#> #   effect_list <list>, mirna_count_per_comp <dbl>