simulate_vis

simulate_vis provides visualisation of the graph in addition to simulate function.

simulate_vis(input_graph, cycle = 1, Competing_color = "green",
  mirna_color = "orange", Upregulation = "red",
  Downregulation = "blue", title = "GRAPH", layout = "kk")

Arguments

input_graph
The graph object that processed in previous steps.

cycle
Optimal iteration number for gaining steady-state.

Competing_color
The color of competing elements on the graph with "green" default.

mirna_color
The color of miRNAs on the graph with "orange" default.

Upregulation
The color of Upregulated elements on the graph with "red" default.

Downregulation
The color of Downregulated elements on the graph with "blue" default.

title
Title of the given graph.

layout
The layout that will be used for visualisation of the graph.

Value

It gives a graph and the images of states in each iteration until the end of the simulation.

Details

simulate_vis gives the last graph object and each iterations' image.

Examples

# When does the system gain steady-state conditions again?

## new_counts, the dataset that includes the current counts of nodes.

priming_graph(minsamp, Competing_expression, miRNA_expression, aff_factor = c(seed_type,energy), deg_factor = c(region))%>%
   update_nodes(once = TRUE)%>%
   update_variables(new_counts)%>%
   update_nodes()%>%
   simulate_vis(cycle = 12)

#> Error in enquo(competing_count): "enquo" fonksiyonu bulunamadı