Carries variables from edge to node.

This function carries variables from edge to node and should be used after `update_how` or `update_variables` functions

update_nodes(input_graph, once = FALSE)

Arguments

input_graph
Processed graph object in previous step.

once
The argument is about when the carrying process runs (internal use only)

Value

the graph object.

Details

If the carrying process performs after priming_graph function, the argument must be TRUE. The function helps to visualisation of processed graph object, especially that includes too many nodes.This step makes it easily to follow the processes.

Examples

data("minsamp")

minsamp %>%
  priming_graph(Gene_expression, miRNA_expression) %>%
  update_how("Gene2",2) %>%
  update_nodes() %>%

#> Error: <text>:10:0: unexpected end of input
#> 8: 
#> 9: 
#>   ^