Calculates average expression changes of all (or specified) nodes except trigger and finds the perturbed node count for all (or specified) nodes in system.

Calculates average expression changes of all (or specified) nodes except trigger and finds the perturbed node count for all (or specified) nodes in system.

find_node_perturbation(input_graph, how = 2, cycle = 1, limit = 0,
  fast = 0)

Arguments

input_graph
The graph object that was processed with priming_graph function.

how
The change of count (expression) of the given node in terms of fold change.

cycle
The iteration of simulation.

limit
The minimum fold change which can be taken into account for perturbation calculation on all nodes in terms of percentage.

included
specifies percentage of affected target in target expression. For example, if included = 1, the nodes that are affected from miRNA repression activity more than one percent of their expression is determined as subgraph.

Value

It gives a tibble form dataset that includes node names, perturbation efficiency and perturbed count of nodes.

Details

find_node_perturbation calculates mean expression changes of elements after the change in the network in terms of percentage. It also calculates the number of nodes that have expression changes after the change occur in the network. The outputs of the function are the perturbation efficiency and perturbed count of nodes for each nodes.

Examples

data("minsamp")
data("midsamp")

 minsamp%>%
  priming_graph(competing_count = Competing_expression, miRNA_count = miRNA_expression)%>%
  find_node_perturbation()

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tibble: 8 x 9
#>   name  type  node_id initial_count count_pre count_current changes_variable
#>   <chr> <chr>   <int>         <dbl>     <dbl>         <dbl> <chr>           
#> 1 Gene1 Comp~       1         10000     10000         10000 Competing       
#> 2 Gene2 Comp~       2         10000     10000         10000 Competing       
#> 3 Gene3 Comp~       3          5000      5000          5000 Competing       
#> 4 Gene4 Comp~       4         10000     10000         10000 Competing       
#> 5 Gene5 Comp~       5          5000      5000          5000 Competing       
#> 6 Gene6 Comp~       6         10000     10000         10000 Competing       
#> 7 Mir1  miRNA       7          1000      1000          1000 miRNA           
#> 8 Mir2  miRNA       8          2000      2000          2000 miRNA           
#> # ... with 2 more variables: perturbation_efficiency <dbl>,
#> #   perturbed_count <dbl>

 minsamp%>%
  priming_graph(competing_count = Competing_expression, miRNA_count = miRNA_expression, aff_factor = c(energy,seed_type), deg_factor = region)%>%
  find_node_perturbation(how = 3, cycle = 4)

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tibble: 8 x 9
#>   name  type  node_id initial_count count_pre count_current changes_variable
#>   <chr> <chr>   <int>         <dbl>     <dbl>         <dbl> <chr>           
#> 1 Gene1 Comp~       1         10000     10000         10000 Competing       
#> 2 Gene2 Comp~       2         10000     10000         10000 Competing       
#> 3 Gene3 Comp~       3          5000      5000          5000 Competing       
#> 4 Gene4 Comp~       4         10000     10000         10000 Competing       
#> 5 Gene5 Comp~       5          5000      5000          5000 Competing       
#> 6 Gene6 Comp~       6         10000     10000         10000 Competing       
#> 7 Mir1  miRNA       7          1000      1000          1000 miRNA           
#> 8 Mir2  miRNA       8          2000      2000          2000 miRNA           
#> # ... with 2 more variables: perturbation_efficiency <dbl>,
#> #   perturbed_count <dbl>

 midsamp%>%
  priming_graph(competing_count = Gene_expression, miRNA_count = miRNA_expression)%>%
  find_node_perturbation(how = 2, cycle= 3, limit=1, fast = 5)

#> Warning: First column is processed as competing and the second as miRNA.

#> Warning: `as_quosure()` requires an explicit environment as of rlang 0.3.0.
#> Please supply `env`.
#> This warning is displayed once per session.

#> Subsetting by edges

#> # A tibble: 15 x 10
#>    name  type  node_id initial_count count_pre count_current changes_variable
#>    <chr> <chr>   <int>         <dbl>     <dbl>         <dbl> <chr>           
#>  1 Gene4 Comp~       4         10000     10000         10000 Competing       
#>  2 Gene6 Comp~       6         10000     10000         10000 Competing       
#>  3 Gene7 Comp~       7          7000      7000          7000 Competing       
#>  4 Gene8 Comp~       8          3000      3000          3000 Competing       
#>  5 Gene9 Comp~       9          6000      6000          6000 Competing       
#>  6 Gene~ Comp~      10          2000      2000          2000 Competing       
#>  7 Gene~ Comp~      11          8000      8000          8000 Competing       
#>  8 Gene~ Comp~      12          1500      1500          1500 Competing       
#>  9 Gene~ Comp~      13           500       500           500 Competing       
#> 10 Gene~ Comp~      14          7000      7000          7000 Competing       
#> 11 Gene~ Comp~      15          3000      3000          3000 Competing       
#> 12 Gene~ Comp~      16          2000      2000          2000 Competing       
#> 13 Gene~ Comp~      17          6000      6000          6000 Competing       
#> 14 Mir3  miRNA      23          3000      3000          3000 miRNA           
#> 15 Mir4  miRNA      24          5000      5000          5000 miRNA           
#> # ... with 3 more variables: .tidygraph_node_index <int>,
#> #   perturbation_efficiency <dbl>, perturbed_count <dbl>