Converts the given dataframe using first variable as competing and the second as miRNA.

The function converts the given dataframe using first variable as competing and the second as miRNA. If user defines interaction factors as affinity or degradation, the factors are taken into account.

priming_graph(df, competing_count, miRNA_count, aff_factor = dummy,
  deg_factor = dummy)

Arguments

df
A data frame that includes the miRNA and competing targets.

competing_count
The counts (or expression) of competing elements of the dataset.

miRNA_count
The counts (or expression) of repressive element (miRNA) of the dataset.

aff_factor
The parameter/s of binding between miRNA and targets.

deg_factor
The parameter/s for degradation of bound miRNA:target complex.

Value

the graph object.

Details

priming_graph provides grouping of competing targets and evaluation of targets within the groups taking into account miRNA:target, target:total target, interaction and degradation parameters. The target groups are determined according to miRNAs. If the factors that are important in target interactions are specified as arguments, the factors also are evaluated separately within each group. priming_graph also calculates the miRNA efficiency in steady-state conditions. It is assumed that quantity of competing targets and miRNAs are shown in the steady-state system after the miRNAs exhibit repressive efficiency. Note that the data must not include missing values such as NA or "-".

Examples

data("minsamp")

priming_graph(minsamp, Competing_expression, miRNA_expression)

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tbl_graph: 8 nodes and 7 edges
#> #
#> # A rooted tree
#> #
#> # Node Data: 8 x 7 (active)
#>   name  type      node_id initial_count count_pre count_current changes_variable
#>   <chr> <chr>       <int>         <dbl>     <dbl>         <dbl> <chr>           
#> 1 Gene1 Competing       1         10000     10000         10000 Competing       
#> 2 Gene2 Competing       2         10000     10000         10000 Competing       
#> 3 Gene3 Competing       3          5000      5000          5000 Competing       
#> 4 Gene4 Competing       4         10000     10000         10000 Competing       
#> 5 Gene5 Competing       5          5000      5000          5000 Competing       
#> 6 Gene6 Competing       6         10000     10000         10000 Competing       
#> # … with 2 more rows
#> #
#> # Edge Data: 7 x 19
#>    from    to Competing_name miRNA_name Competing_expre… miRNA_expression dummy
#>   <int> <int> <chr>          <chr>                 <dbl>            <dbl> <dbl>
#> 1     1     7 Gene1          Mir1                  10000             1000     1
#> 2     2     7 Gene2          Mir1                  10000             1000     1
#> 3     3     7 Gene3          Mir1                   5000             1000     1
#> # … with 4 more rows, and 12 more variables: afff_factor <dbl>,
#> #   degg_factor <dbl>, comp_count_list <list>, comp_count_pre <dbl>,
#> #   comp_count_current <dbl>, mirna_count_list <list>, mirna_count_pre <dbl>,
#> #   mirna_count_current <dbl>, mirna_count_per_dep <dbl>, effect_current <dbl>,
#> #   effect_pre <dbl>, effect_list <list>

priming_graph(minsamp, Competing_expression, miRNA_expression, aff_factor = c(seed_type,energy), deg_factor = region)

#> Warning: First column is processed as competing and the second as miRNA.

#> # A tbl_graph: 8 nodes and 7 edges
#> #
#> # A rooted tree
#> #
#> # Node Data: 8 x 7 (active)
#>   name  type      node_id initial_count count_pre count_current changes_variable
#>   <chr> <chr>       <int>         <dbl>     <dbl>         <dbl> <chr>           
#> 1 Gene1 Competing       1         10000     10000         10000 Competing       
#> 2 Gene2 Competing       2         10000     10000         10000 Competing       
#> 3 Gene3 Competing       3          5000      5000          5000 Competing       
#> 4 Gene4 Competing       4         10000     10000         10000 Competing       
#> 5 Gene5 Competing       5          5000      5000          5000 Competing       
#> 6 Gene6 Competing       6         10000     10000         10000 Competing       
#> # … with 2 more rows
#> #
#> # Edge Data: 7 x 22
#>    from    to Competing_name miRNA_name Competing_expre… miRNA_expression
#>   <int> <int> <chr>          <chr>                 <dbl>            <dbl>
#> 1     1     7 Gene1          Mir1                  10000             1000
#> 2     2     7 Gene2          Mir1                  10000             1000
#> 3     3     7 Gene3          Mir1                   5000             1000
#> # … with 4 more rows, and 16 more variables: seed_type <dbl>, energy <dbl>,
#> #   region <dbl>, dummy <dbl>, afff_factor <dbl>, degg_factor <dbl>,
#> #   comp_count_list <list>, comp_count_pre <dbl>, comp_count_current <dbl>,
#> #   mirna_count_list <list>, mirna_count_pre <dbl>, mirna_count_current <dbl>,
#> #   mirna_count_per_dep <dbl>, effect_current <dbl>, effect_pre <dbl>,
#> #   effect_list <list>